Cleanup renderers.
