+++ src/basestyle/index.less
@@ -12,4 +12,5 @@
 @import '../filebrowser/base.less';
 @import '../iframe/base.less';
 @import '../notebook/base.less';
+@import '../renderers/base.less';
 @import '../running/base.less';
