+++ src/default-theme/index.less
@@ -24,6 +24,6 @@
 @import './mainmenu.less';
 @import './markdownwidget.less';
 @import './notebook.less';
-@import './rendermime.less';
+@import './renderers.less';
 @import './running.less';
 @import './terminal.less';
