+++ src/default-theme/renderers.less
@@ -1,0 +1,13 @@
+/*-----------------------------------------------------------------------------
+| Copyright (c) Jupyter Development Team.
+| Distributed under the terms of the Modified BSD License.
+|----------------------------------------------------------------------------*/
+
+
+.jp-RenderedHTMLCommon {
+
+  blockquote {
+    border-left: 5px solid #eeeeee;
+  }
+
+}
