+++ src/notebook/base.less
@@ -3,6 +3,8 @@
 |
 | Distributed under the terms of the Modified BSD License.
 |----------------------------------------------------------------------------*/
+
+
 .jp-Notebook-panel {
   display: flex;
   flex-direction: column;
@@ -38,68 +40,7 @@
 
 
 .jp-MarkdownCell-content {
-  padding: 7px;
-  user-select: text;
-  -moz-user-select: text;
-  -webkit-user-select: text;
-  -ms-user-select: text;
-}
-
-
-.jp-MarkdownCell-content h1 {
-  font-size: 26px;
-  line-height: 1.0;
-  margin: 14px 0 0 0;
-}
-
-
-.jp-MarkdownCell-content h2 {
-  font-size: 22px;
-  line-height: 1.0;
-  margin: 14px 0 0 0;
-}
-
-
-.jp-MarkdownCell-content h3 {
-  font-size: 18px;
-  line-height: 1.0;
-  margin: 14px 0 0 0;
-}
-
-
-.jp-MarkdownCell-content h4 {
-  font-size: 14px;
-  line-height: 1.0;
-  margin: 14px 0 0 0;
-}
-
-
-.jp-MarkdownCell-content h5 {
-  font-size: 14px;
-  font-style: italic;
-  line-height: 1.0;
-  margin: 14px 0 0 0;
-}
-
-
-.jp-MarkdownCell-content h6 {
-  font-size: 14px;
-  font-style: italic;
-  line-height: 1.0;
-  margin: 14px 0 0 0;
-}
-
-
-.jp-MarkdownCell-content p {
-  font-size: 14px;
-  line-height: 20px;
-  margin: 0 0 0 0;
-}
-
-
-.jp-MarkdownCell-content li {
-  line-height: 20px;
-  margin: 14px 0 0 0;
+  padding: 0.5em;
 }
 
 
@@ -110,7 +51,7 @@
   -ms-user-select: text;
 }
 
-.jp-Output-result {
+.p-Widget.jp-Output-result {
   overflow: auto;
 }
