+++ src/renderers/base.less
@@ -1,0 +1,291 @@
+/*-----------------------------------------------------------------------------
+| Copyright (c) Jupyter Development Team.
+| Distributed under the terms of the Modified BSD License.
+|----------------------------------------------------------------------------*/
+
+.jp-RenderedHTMLCommon {
+
+  color: black;
+  font-size: 14px;
+  line-height: 20px;
+
+  .MathJax_Display {
+    margin: 0;
+  }
+
+
+  em {
+    font-style: italic;
+  }
+
+  strong {
+    font-weight: bold;
+  }
+
+  u {
+    text-decoration: underline;
+  }
+
+  :link {
+    text-decoration: underline;
+  }
+
+  :visited {
+    text-decoration: underline;
+  }
+
+  /*For a 14px base font size this goes as:
+  font-size = 26, 22, 18, 14, 12, 12
+  margin-top = 14, 14, 14, 14, 8, 8
+  */
+
+  h1 {
+    font-size: 185.7%;
+    margin: 1.08em 0 0 0;
+    font-weight: bold;
+    line-height: 1.0;
+  }
+
+
+  h2 {
+    font-size: 157.1%;
+    margin: 1.27em 0 0 0;
+    font-weight: bold;
+    line-height: 1.0;
+  }
+
+
+  h3 {
+    font-size: 128.6%;
+    margin: 1.55em 0 0 0;
+    font-weight: bold;
+    line-height: 1.0;
+  }
+
+
+  h4 {
+    font-size: 100%;
+    margin: 2em 0 0 0;
+    font-weight: bold;
+    line-height: 1.0;
+  }
+
+
+  h5 {
+    font-size: 100%;
+    margin: 2em 0 0 0;
+    font-weight: bold;
+    line-height: 1.0;
+    font-style: italic;
+  }
+
+
+  h6 {
+    font-size: 100%;
+    margin: 2em 0 0 0;
+    font-weight: bold;
+    line-height: 1.0;
+    font-style: italic;
+  }
+
+
+  h1:first-child {
+    margin-top: 0.538em;
+  }
+
+
+  h2:first-child {
+    margin-top: 0.636em;
+  }
+
+
+  h3:first-child {
+    margin-top: 0.777em;
+  }
+
+
+  h4:first-child {
+    margin-top: 1em;
+  }
+
+
+  h5:first-child {
+    margin-top: 1em;
+  }
+
+
+  h6:first-child {
+    margin-top: 1em;
+  }
+
+
+  ul:not(.list-inline),
+  ol:not(.list-inline) {
+    padding-left: 2em;
+  }
+
+
+  ul {
+    list-style: disc;
+  }
+
+
+  ul ul {
+    list-style: square;
+  }
+
+
+  ul ul ul {
+    list-style: circle;
+  }
+
+
+  ol {
+    list-style: decimal;
+  }
+
+
+  ol ol {
+    list-style: upper-alpha;
+  }
+
+
+  ol ol ol {
+    list-style: lower-alpha;
+  }
+
+
+  ol ol ol ol {
+    list-style: lower-roman;
+  }
+
+
+  ol ol ol ol ol {
+    list-style: decimal;
+  }
+
+
+  * + ul {
+    margin-top: 1em;
+  }
+
+
+  * + ol {
+    margin-top: 1em;
+  }
+
+
+  hr {
+    color: black;
+    background-color: black;
+    margin-top: 1em;
+    margin-bottom: 1em;
+  }
+
+
+  pre {
+    margin: 1em 2em;
+  }
+
+
+  pre,
+  code {
+    border: 0;
+    background-color: white;
+    color: black;
+    font-size: 100%;
+    padding: 0px;
+  }
+
+
+  blockquote {
+    margin: 1em 2em;
+  }
+
+
+  table {
+    margin-left: auto;
+    margin-right: auto;
+    border: 1px solid black;
+    border-collapse: collapse;
+  }
+
+
+  tr,
+  th,
+  td {
+    border: 1px solid black;
+    border-collapse: collapse;
+    margin: 1em 2em;
+  }
+
+
+  td,
+  th {
+    text-align: left;
+    vertical-align: middle;
+    padding: 4px;
+  }
+
+
+  th {
+    font-weight: bold;
+  }
+
+  * + table {
+    margin-top: 1em;
+  }
+
+
+  p {
+    text-align: left;
+    margin: 0px;
+  }
+
+
+  * + p {
+    margin-top: 1em;
+  }
+
+
+  img {
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+  }
+
+
+  * + img {
+    margin-top: 1em;
+  }
+
+
+  img,
+  svg {
+    max-width: 100%;
+    height: auto;
+  }
+
+
+  img.unconfined,
+  svg.unconfined {
+    max-width: none;
+  }
+
+  .alert {
+    margin-bottom: initial;
+  }
+
+
+  * + .alert {
+    margin-top: 1em;
+  }
+
+
+  blockquote {
+    margin: 1em 2em;
+    padding: 0 1em;
+    border-left: 5px solid black;
+  }
+
+}
